feat: Set window z-order to bottom in MidasWindowManager
This commit modifies the MidasWindowManager class in the app.py file to set the window z-order to the bottom. It adds the win32gui.SetWindowPos function with the win32con.HWND_BOTTOM parameter to ensure that the window is always positioned at the bottom of the z-order. This improves the visual hierarchy and ensures that the Midas window is not obstructed by other windows.